Move context state to context.py
