+++ python/ray/serve/constants.py
@@ -1,4 +1,5 @@
 from enum import Enum
+import re
 
 #: Used for debugging to turn on DEBUG-level logs
 DEBUG_LOG_ENV_VAR = "SERVE_DEBUG_LOG"
@@ -86,6 +87,14 @@ REPLICA_HEALTH_CHECK_UNHEALTHY_THRESHOLD = 3
 # Key used to idenfity given json represents a serialized RayServeHandle
 SERVE_HANDLE_JSON_KEY = "__SerializedServeHandle__"
 
+# The time in seconds that the Serve client waits before rechecking deployment state
+CLIENT_POLLING_INTERVAL_S: float = 1
+
+# Regex pattern for anonymous namespace. Should match the pattern used in
+# src/ray/gcs/gcs_server/gcs_actor_manager.cc's is_uuid() method.
+ANONYMOUS_NAMESPACE_PATTERN = re.compile(
+    "[a-f0-9]{8}-[a-f0-9]{4}-4[a-f0-9]{3}-[89aAbB][a-f0-9]{3}-[a-f0-9]{12}"
+)
 
 class ServeHandleType(str, Enum):
     SYNC = "SYNC"
